hack,Makefile: Enable tech preview features after cluster creation
Adds a hack script that's responsible for enabling TP features after
cluster creation, and ensuring that the PO components roll out
correctly. Note: this was copied from the CAPI component's
implementation.

This should be a short term hack as the CI workflows
that enable tech preview features before cluster creation
are currently permafailing.

Signed-off-by: timflannagan <timflannagan@gmail.com>