test/e2e: Use foreground cascading deletion
When deleting any PlatformOperator or BundleDeployment resources,
specify the foreground cascading deletion propagation policy in relevant
client usage in the e2e suite.

This should cut down on the number of flakes that result in failed
installations due to tests re-using the same PlatformOperator package
name. The issue here is we don't give the underlying cascading deletion
enough time to fully uninstall before we attempt a new installation.
And the result here is that the BundleDeployment installation fails as
some resources already exist and cannot be successfully adopted.

Signed-off-by: timflannagan <timflannagan@gmail.com>