Temporarily remove from payload
Temporarily remove this image from the release payload. We are moving
the rukpak manifests to openshift/cluster-olm-operator and have to do it
through a series of PRs. The easiest way to do the move is to stop
including platform-operators in the payload, then add it back after the
move is done.

Signed-off-by: Andy Goldstein <andy.goldstein@redhat.com>