manifests: Simplify filenames for default runlevel
0000_ prefixing to explicitly set a filename is a power move for
operators that want to break out of the default [1].  Using that
mechanism to declare "I'm fine with the default 50 runlevel" does not
change your runlevel, and makes the filenames longer in this
repository.  In this commit, I'm dropping that prefix and now only
explicitly talking about the relative order between this repository's
manifests [2], while deferring cross-repository considerations like
runlevels to 'oc adm release new ...'.

[1]: https://github.com/openshift/enhancements/blob/43ae3ee79a5690642a4e0540c07fb558e322d88e/dev-guide/cluster-version-operator/dev/operators.md#how-do-i-get-added-as-a-special-run-level
[2]: https://github.com/openshift/enhancements/blob/43ae3ee79a5690642a4e0540c07fb558e322d88e/dev-guide/cluster-version-operator/dev/operators.md#names-of-manifest-files